@@ -5,11 +5,13 @@
 namespace denha;
 
 use denha\Config;
+use denha\Start;
 
 class Trace
 {
     private static $tracePageTabs  = ['BASE' => '基本', 'FILE' => '文件', 'ERR|NOTIC' => '错误', 'SQL' => 'SQL', 'DEBUG' => '调试'];
     private static $traceErrorType = [0 => '', 1 => 'FATAL ERROR', 2 => 'WARNING', 4 => 'PARSE', 8 => 'NOTICE', 100 => 'SQL'];
+    private static $dbAllTime      = 0; // 数据运行时间
 
     public static $errorInfo = []; //错误信息
     public static $sqlInfo   = []; //sql执行信息
@@ -56,6 +58,7 @@ class Trace
     //捕获Notice错误信息
     public static function catchNotice($level, $message, $file, $line)
     {
+
         if ($level) {
 
             $debugContent = self::getFileContent($file, $line);
@@ -71,6 +74,7 @@ class Trace
                     'line'    => $line,
                 );
 
+                header('HTTP/1.1 500 Internal Server Error');
                 return include FARM_PATH . DS . 'trace' . DS . 'error.html';
             }
         }
@@ -79,6 +83,7 @@ class Trace
     // 捕获致命错误信息 并显示
     public static function catchError()
     {
+
         $e = error_get_last();
         if ($e) {
 
@@ -104,15 +109,9 @@ class Trace
             }
 
             if (Config::get('debug')) {
+                header('HTTP/1.1 500 Internal Server Error');
                 return include FARM_PATH . DS . 'trace' . DS . 'error.html';
             } else {
-                //FATAL ERROR 发送到邮箱
-                // if (Config::get('send_debug_mail')) {
-                //     $title   = $_SERVER['HTTP_HOST'] . ' 有一个致命错误 ip:' . getIP() . ' ' . $_SERVER['SERVER_PROTOCOL'];
-                //     $content = $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'] . PHP_EOL . 'FATAL ERROR : ' . $e['message'] . ' from ' . $e['file'] . ' on line ' . $e['line'];
-                //     dao('Mail')->send(Config::get('send_mail'), $title, $content);
-                // }
-
                 header("http/1.1 404 not found");
                 header("status: 404 not found");
                 return include FARM_PATH . DS . 'trace' . DS . '404.html';
@@ -123,6 +122,7 @@ class Trace
     //捕获未处理的自定义错误信息 并显示
     public static function catchApp($error)
     {
+
         $e['type']    = 0;
         $e['message'] = $error->getMessage();
         $e['file']    = $error->getFile();
@@ -147,6 +147,7 @@ class Trace
         }
 
         if (Config::get('trace')) {
+            header('HTTP/1.1 500 Internal Server Error');
             return include FARM_PATH . DS . 'trace' . DS . 'error.html';
         } else {
             header("http/1.1 404 not found");
@@ -176,6 +177,8 @@ class Trace
                     $info[] = 'EXPLAIN :' . json_encode($explain);
                 }
             }
+
+            self::$dbAllTime += $data['time'];
         } else {
             $info[] = $data;
         }
@@ -226,16 +229,17 @@ class Trace
         }
 
         $base = [
-            '请求信息' => date('Y-m-d H:i:s', $_SERVER['REQUEST_TIME']) . ' ' . $_SERVER['SERVER_PROTOCOL'] . ' ' . $_SERVER['REQUEST_METHOD'] . ' : ' . strip_tags($_SERVER['REQUEST_URI']),
-            '运行时间' => number_format(microtime(true) - START_TIME, 6) . ' s',
-            '吞吐率'    => number_format(1 / (microtime(true) - START_TIME), 2) . 'req/s',
-            '内存开销' => number_format((memory_get_usage() - START_USE_MENUS) / 1024, 2) . ' kb',
-            '文件加载' => count(get_included_files()),
-            //'缓存信息' => n('cache_read') . ' gets ' . n('cache_write') . ' writes ',
-            '配置加载' => count(config()),
-            '会话信息' => 'SESSION_ID=' . session_id(),
-            '数据库'    => $dbName,
-            '磁盘信息' => number_format(DISK_TOTAL_SPACE / 1024 / 1024 / 1024, 3) . ' G (all) / ' . number_format((DISK_TOTAL_SPACE - DISK_FREE_SPACE) / 1024 / 1024 / 1024, 3) . ' G (use) / ' . number_format(DISK_FREE_SPACE / 1024 / 1024 / 1024, 3) . 'G (free)',
+            '请求信息'          => date('Y-m-d H:i:s', $_SERVER['REQUEST_TIME']) . ' ' . $_SERVER['SERVER_PROTOCOL'] . ' ' . $_SERVER['REQUEST_METHOD'] . ' : ' . strip_tags($_SERVER['REQUEST_URI']),
+            '运行时间'          => number_format(microtime(true) - START_TIME, 6) . ' s',
+            '吞吐率'             => number_format(1 / (microtime(true) - START_TIME), 2) . 'req/s',
+            '内存开销'          => number_format((memory_get_usage() - START_USE_MENUS) / 1024, 2) . ' kb',
+            '文件加载'          => count(get_included_files()),
+            '配置加载'          => count(Start::$config),
+            '会话信息'          => 'SESSION_ID=' . session_id(),
+
+            '数据库运行时间' => self::$dbAllTime,
+            '数据库'             => $dbName,
+            '磁盘信息'          => number_format(DISK_TOTAL_SPACE / 1024 / 1024 / 1024, 3) . ' G (all) / ' . number_format((DISK_TOTAL_SPACE - DISK_FREE_SPACE) / 1024 / 1024 / 1024, 3) . ' G (use) / ' . number_format(DISK_FREE_SPACE / 1024 / 1024 / 1024, 3) . 'G (free)',
         ];
 
         return $base;
